backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pyjwt==2.9.0
pycryptodome==3.20.0
python-dotenv==1.0.1
python-jwt==4.1.0  
bcrypt==4.0.1  
cryptography==41.0.4  
slowapi==0.1.8  
pydantic==2.4.2  

